moved searchPlaces to a new file
